Expand INT() to include unsigned types so u8 <: int; add supertype annotation test

+++ lentoc/lib/src/type_checker/types.rs
@@ -824,7 +824,10 @@ pub mod std_types {
     pub fn INT() -> Type {
         Type::Alias(
             Str::Str("int"),
-            Box::new(Type::Sum(vec![INT8, INT16, INT32, INT64, INT128, INTBIG])),
+            Box::new(Type::Sum(vec![
+                UINT(),
+                INT8, INT16, INT32, INT64, INT128, INTBIG,
+            ])),
         )
     }
